Introduced game quality control game_qc introduced to agent.py Algorithm: the game is selected with a score >= a certain quotient of the highest score thus far, or the highest score game of a certain number.
Example: let us say the highest game score attained thus far is 10,000.
We then execute a number of games saving the one with the highest score in
Let us say the quality threshold is set at 0.8 and the maximum replay limit
is set at 100. That means we select either the game with a score of
0.8 * 10,000 = 8,000 or above, or we play 100 games and select the one with
the highest score out of that sequence.

Documentation introduced throughout the system